+++ NEWS.md
@@ -1,6 +1,6 @@
 # usethis (development version)
 
-`use_github_actions_badge()` now uses the same URLs as GitHub does via the "Create status badge" helper in the browser (#1525). This changes the signicance of the `name` argument; now it really must be the name of the workflow configuration file.
+`use_github_actions_badge()` now uses the same URLs as GitHub does via the "Create status badge" helper in the browser (#1525). This changes the significance of the `name` argument; now it really must be the name of the workflow configuration file.
 
 # usethis 2.1.5