+++ NEWS.md
@@ -1,4 +1,4 @@
-# usethis (development version)
+# usethis 2.1.6
 
 ### GitHub-related